pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cogito-ergo"
version = "0.3.0"
description = "Memory retrieval for AI agents — integer-pointer fidelity guarantee, 93.4% R@1 on LongMemEval_S (hybrid tier). By Hermes Labs."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
  { name = "Hermes Labs", email = "roli@hermes-labs.ai" },
]
keywords = [
  "ai", "agents", "memory", "rag", "retrieval", "llm",
  "mem0", "agent", "zer0dex", "integer-pointer", "fidelity",
  "vector-store", "chromadb", "ollama",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
]
dependencies = [
  "mem0ai>=1.0.5",
  "chromadb>=0.5.0",
]

[project.optional-dependencies]
ollama = []  # no extra deps — Ollama runs locally
hybrid = ["bm25s>=0.2"]  # BM25 fusion for recall_hybrid; degrades gracefully if absent
dev = ["pytest", "pytest-asyncio", "httpx", "ruff"]

[project.scripts]
cogito = "cogito.cli:main"
cogito-server = "cogito.server:main"
cogito-scaffold = "cogito.scaffold_server:main"

[project.urls]
Homepage = "https://hermes-labs.ai"
Repository = "https://github.com/roli-lpci/cogito-ergo"
Documentation = "https://github.com/roli-lpci/cogito-ergo#readme"
Issues = "https://github.com/roli-lpci/cogito-ergo/issues"
Changelog = "https://github.com/roli-lpci/cogito-ergo/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/cogito"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
